requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
click>=8.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart
python-dotenv>=1.0.0
httpx
aiohttp
aiofiles
numpy
cryptography>=3.4.8
pyjwt>=2.10.1
python-jose[cryptography]>=3.5.0
redis
sqlalchemy
asyncpg
supabase>=2.9.0
postgrest>=0.16.11
storage3>=0.7.7
gotrue>=2.8.4
realtime>=2.0.2
supafunc>=0.4.6
websockets>=12.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.8.5
pypdf>=6.0.0
faiss-cpu>=1.12.0
rank_bm25
typing-extensions>=4.5.0
anyio>=3.0.0
sniffio>=1.1.0
h11>=0.8
watchfiles>=0.13
python-multipart>=0.0.5
